transaction service reversepeyment

+++ Entregables/Java_II/tc/src/main/java/com/parquesoftti/tc/service/impl/TransactionServiceImpl.java
@@ -9,6 +9,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Optional;
+
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
 @Service
 public class TransactionServiceImpl implements TransactionService {
@@ -26,8 +28,15 @@ public class TransactionServiceImpl implements TransactionService {
 
     @Transactional(readOnly = false, rollbackFor = Exception.class, propagation = Propagation.REQUIRED)
     @Override
-    public void deleteCard(Long id) {
+    public void reversePayment(Long id) {
+        if (id == null || id <=0) {
+            throw new IllegalArgumentException("El Id debe ser positivo y no nulo.");
+        }
 
+        Optional<Transaction> existingTransaction = transactionRepository.findById(id);
+        if (existingTransaction.isEmpty()) {
+            throw new RuntimeException("La transacción no existe y no se puede eliminar.");
+        }
         transactionRepository.deleteById(id);
     }
